refactor(core): split geometry marks and extract finalizePipelineRun
Collapse remaining post-scale orchestration into finalize-run, and move
point/path, bar/rule/text, and composite geom builders into focused modules
behind thin facades. Characterization tests lock layer-major paint order and
geom batch kinds.
