+++ packages/core/src/pipeline.ts
@@ -39,22 +39,12 @@
 import type { SpecInput, PortableSpec } from "@ggsvelte/spec";
 
 import { perfMark, perfMeasure } from "./perf.js";
-import { LineageStore } from "./identity.js";
 
 import type { Advisory, PipelineWarning, RenderModel, RunOptions } from "./pipeline/types.js";
-import { preparePanels } from "./pipeline/prepare-panels.js";
 import { setupPipelineRun } from "./pipeline/setup-run.js";
+import { preparePanels } from "./pipeline/prepare-panels.js";
 import { trainPipelineScales } from "./pipeline/train-pipeline-scales.js";
-import { computePanelLayout } from "./pipeline/panel-layout.js";
-import { assembleScene, buildGeometryBatches } from "./pipeline/assemble-scene.js";
-import {
-  resolveLayerBackends,
-  resolveLayerFields,
-  resolveLayerScaledConstants,
-} from "./pipeline/layer-contracts.js";
-import { buildPipelineCandidates } from "./pipeline/build-candidates.js";
-import { computeBaselineDomains, computeEffectiveDomains } from "./pipeline/compute-domains.js";
-import { assembleRenderModel } from "./pipeline/assemble-render-model.js";
+import { finalizePipelineRun } from "./pipeline/finalize-run.js";
 
 // Re-export the public pipeline contract (import path stability).
 export type {
@@ -94,32 +84,21 @@ export function runPipeline(spec: SpecInput | PortableSpec, options: RunOptions)
 
   // bind + facet partition + per-panel frames
   perfMark("ggsvelte:bind:start");
-  const { table, faceted, freeX, freeY, nrow, ncol, facetPanels, bindings, panelFrames } =
-    preparePanels(normalized, options, warnings, advisories);
+  const prepared = preparePanels(normalized, options, warnings, advisories);
   perfMark("ggsvelte:bind:end");
   perfMeasure("ggsvelte:bind", "ggsvelte:bind:start", "ggsvelte:bind:end");
 
   // train scales — fixed: union across panels; free: positional domains per
   // panel; discrete color/fill assignment ALWAYS global (one legend).
   perfMark("ggsvelte:scales:start");
-  const {
-    xTraining,
-    yTraining,
-    panelScales,
-    colorResolution,
-    fillResolution,
-    xInputs,
-    yInputs,
-    scalesConfig,
-    allFrames,
-  } = trainPipelineScales({
+  const trained = trainPipelineScales({
     normalized,
     options,
-    table,
-    facetPanels,
-    panelFrames,
-    freeX,
-    freeY,
+    table: prepared.table,
+    facetPanels: prepared.facetPanels,
+    panelFrames: prepared.panelFrames,
+    freeX: prepared.freeX,
+    freeY: prepared.freeY,
     editionDefaults,
     warnings,
     advisories,
@@ -127,129 +106,19 @@ export function runPipeline(spec: SpecInput | PortableSpec, options: RunOptions)
   perfMark("ggsvelte:scales:end");
   perfMeasure("ggsvelte:scales", "ggsvelte:scales:start", "ggsvelte:scales:end");
 
-  // layout (bounded two-pass; facet grids run the per-panel mirror of it)
-  perfMark("ggsvelte:layout:start");
-  const panelLayout = computePanelLayout({
-    flip,
-    faceted,
-    freeX,
-    freeY,
-    nrow,
-    ncol,
-    facetPanels,
-    panelScales,
-    allFrames,
-    labs: normalized.labs ?? {},
-    scalesConfig,
-    xScale: xTraining.scale,
-    yScale: yTraining.scale,
-    colorLegend: colorResolution.legendInput,
-    fillLegend: fillResolution.legendInput,
-    legendOrder: normalized.legend?.order ?? "stable-domain",
-    theme,
-    options,
-    warnings,
-  });
-  perfMark("ggsvelte:layout:end");
-  perfMeasure("ggsvelte:layout", "ggsvelte:layout:start", "ggsvelte:layout:end");
-
-  // geometry (panel-local px; coord flip transforms per batch).
-  // LAYER-major order: layer order is paint order across the whole plot, and
-  // it keeps each layer's batches contiguous so strata planning (contiguous
-  // same-backend batches share a stratum) never fragments on facet panels.
-  perfMark("ggsvelte:geometry:start");
-  const batches = buildGeometryBatches({
-    layerCount: normalized.layers.length,
-    facetPanels,
-    panelFrames,
-    placements: panelLayout.placements,
-    panelScales,
-    color: colorResolution.resolved,
-    fill: fillResolution.resolved,
-    flip,
-    warnings,
-  });
-  perfMark("ggsvelte:geometry:end");
-  perfMeasure("ggsvelte:geometry", "ggsvelte:geometry:start", "ggsvelte:geometry:end");
-
-  const scene = assembleScene({
-    width: options.width,
-    height: options.height,
-    placements: panelLayout.placements,
-    facetPanels,
-    displayScales: panelLayout.displayScales,
-    hTitle: panelLayout.hTitle,
-    vTitle: panelLayout.vTitle,
-    batches,
-    legendBlock: panelLayout.legendBlock,
-    topBand: panelLayout.topBand,
-    theme,
-    title: panelLayout.title,
-    subtitle: panelLayout.subtitle,
-    caption: panelLayout.caption,
-  });
-
-  const layerBackends = resolveLayerBackends(
-    normalized.layers,
-    batches,
-    normalized.a11y,
-    options.canvasThreshold,
-    advisories,
-  );
-  const layerFields = resolveLayerFields(normalized.layers.length, bindings);
-  const layerScaledConstants = resolveLayerScaledConstants(normalized.layers.length, bindings);
-
-  const effectiveDomains = computeEffectiveDomains(xTraining.scale, yTraining.scale, panelScales);
-  const baselineDomains = computeBaselineDomains({
-    options,
-    freeX,
-    freeY,
-    facetPanels,
-    panelFrames,
-    xInputs,
-    yInputs,
-    effectiveDomains,
-  });
-
-  const lineage = new LineageStore<number>();
-  const candidates = buildPipelineCandidates({
-    scene,
+  const model = finalizePipelineRun({
     runId,
+    normalized,
+    options,
+    theme,
     flip,
-    bindings,
-    panelFrames,
-    facetPanels,
-    table,
-    layerFields,
-    color: colorResolution.resolved,
-    fill: fillResolution.resolved,
-    lineage,
+    prepared,
+    trained,
+    warnings,
+    advisories,
   });
 
   perfMark("ggsvelte:pipeline:end");
   perfMeasure("ggsvelte:pipeline", "ggsvelte:pipeline:start", "ggsvelte:pipeline:end");
-
-  return assembleRenderModel({
-    scene,
-    xScale: xTraining.scale,
-    yScale: yTraining.scale,
-    color: colorResolution.resolved,
-    fill: fillResolution.resolved,
-    panelScales,
-    colorState: colorResolution.state,
-    fillState: fillResolution.state,
-    warnings,
-    advisories,
-    runId,
-    layerBackends,
-    layerFields,
-    layerScaledConstants,
-    baselineDomains,
-    effectiveDomains,
-    lineage,
-    candidates,
-    formatX: panelLayout.formatX,
-    formatY: panelLayout.formatY,
-    table,
-  });
+  return model;
 }
